+++ src/types/modal.ts
@@ -1,0 +1,20 @@
+import type { DeckData } from './deck';
+
+// モーダル関連の型定義
+
+export interface ModalImage {
+  url: string;
+  deckCode: string;
+  playerName?: string;
+  index: number;
+}
+
+export interface ModalState {
+  enlargedImage: ModalImage | null;
+}
+
+export interface ModalActions {
+  openModal: (deck: DeckData, index: number) => void;
+  closeModal: () => void;
+  navigateModal: (direction: "prev" | "next") => void;
+}
